Pom.xml updated, Abstract entity removed, Equation Overhaul required
Equation needs a significant overhaul, including changes in the
variables, the annotations need to reflect the actual content, and the
overall equation needs to be changed due to new information. Some of the
current variables will need to be changed, and entire file may be nuked
in favor of rewriting.